Modify mock match function to return an exact number of matches (#23)
## Description

This PR changes `genFakeMatchData` so that it accepts an integer which is interpreted as the number of mock matches to return.

Also reduces the number of matches showing up in `home` view to 2 matches, and in the `find` view to 5 matches so as to avoid repeated match entries. This is in preparation for match filtering/sorting functionality I'm going to open in a followup PR - having repeated matches show up will look odd especially once filters are applied.